DR, REFLECT, MID DASH, VANISH, SUPER, MULTI HIT FIX

+++ Cube Test 2/Assets/Scripts/Character Scripts/CharacterStateController.cs
@@ -42,7 +42,7 @@ namespace CharacterControl
 
         private float healthPoints = 10000;
 
-        private int superBar = 0;
+        private float superBar = 0;
 
         private Enums.Inputs lastInput;
